requirements.txt
nltk==3.8.1
Pillow==9.5.0
Requests==2.31.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.1
